+++ counterpartycli/server.py
@@ -22,7 +22,6 @@ CONFIG_ARGS = [
     [('--testnet',), {'action': 'store_true', 'default': False, 'help': 'use {} testnet addresses and block numbers'.format(config.BTC_NAME)}],
     [('--testcoin',), {'action': 'store_true', 'default': False, 'help': 'use the test {} network on every blockchain'.format(config.XCP_NAME)}],
 
-    [('--backend-name',), {'default': 'addrindex', 'help': 'the backend name to connect to'}],
     [('--backend-connect',), {'default': 'localhost', 'help': 'the hostname or IP of the backend server'}],
     [('--backend-port',), {'type': int, 'help': 'the backend port to connect to'}],
     [('--backend-user',), {'default': 'bitcoinrpc', 'help': 'the username used to communicate with backend'}],
@@ -32,6 +31,9 @@ CONFIG_ARGS = [
     [('--backend-poll-interval',), {'type': float, 'default': 0.5, 'help': 'poll interval, in seconds (default: 0.5)'}],
     [('--no-check-asset-conservation',), {'action': 'store_true', 'default': False, 'help': 'Skip asset conservation checking (default: false)'}],
     [('--p2sh-dust-return-pubkey',), {'help': 'pubkey to receive dust when multisig encoding is used for P2SH source (default: pubkey from counterparty foundation)'}],
+
+    [('--indexd-connect',), {'default': 'localhost', 'help': 'the hostname or IP of the indexd server'}],
+    [('--indexd-port',), {'type': int, 'default': 8434, 'help': 'the indexd server port to connect to'}],
 
     [('--rpc-host',), {'default': 'localhost', 'help': 'the IP of the interface to bind to for providing JSON-RPC API access (0.0.0.0 for all interfaces)'}],
     [('--rpc-port',), {'type': int, 'help': 'port on which to provide the {} JSON-RPC API'.format(config.APP_NAME)}],
@@ -118,7 +120,6 @@ def main():
         init_args = dict(database_file=args.database_file,
                                 log_file=args.log_file, api_log_file=args.api_log_file,
                                 testnet=args.testnet, testcoin=args.testcoin,
-                                backend_name=args.backend_name,
                                 backend_connect=args.backend_connect,
                                 backend_port=args.backend_port,
                                 backend_user=args.backend_user,
@@ -126,6 +127,7 @@ def main():
                                 backend_ssl=args.backend_ssl,
                                 backend_ssl_no_verify=args.backend_ssl_no_verify,
                                 backend_poll_interval=args.backend_poll_interval,
+                                indexd_connect=args.indexd_connect, indexd_port=args.indexd_port,
                                 rpc_host=args.rpc_host, rpc_port=args.rpc_port, rpc_user=args.rpc_user,
                                 rpc_password=args.rpc_password, rpc_no_allow_cors=args.rpc_no_allow_cors,
                                 requests_timeout=args.requests_timeout,